test: scope impl-specific float conversions to llgo
